+++ typescript/src/index.ts
@@ -2,8 +2,6 @@ export { KURTOSIS_ENGINE_API_VERSION } from "./kurtosis_engine_api_version/kurto
 export { LISTEN_PORT, LISTEN_PROTOCOL } from "./kurtosis_engine_rpc_api_consts/kurtosis_engine_rpc_api_consts";
 
 export { KurtosisContext } from "./lib/kurtosis_context/kurtosis_context";
-export { EnclaveContext } from "./lib/enclave_context/enclave_context";
-export { ApiContainerContext } from "./lib/api_container_context/api_container_context"
 
 // RPC API bindings
 export { EngineServiceClient } from "./kurtosis_engine_rpc_api_bindings/engine_service_grpc_pb";